changed tests for signup and signin to pytest
